Update parse_syscall_data fn to parallelize regex
The first implementation ran the regex line by
line, inserting into BTreeMaps as it went.
This commit changes to run the regex in parallel
and collect into an intermediate Vec. The Vec
is then iterated through sequentially to build
the data maps. This is much faster, as the regex
is the most expensive operation by far.

Also switched data from BTreeMaps to HashMaps,
and updated get_sorted_pid_summaries to allow
results to be sorted by pid, active_time, or
total_time.