Add view of slowest file accesses to PID details
Replaced the list of all files opened with a list
of the ten slowest file opens on PID details.
This is only done on PID details because it is
an expensive operation. We need to split the
lines for each PID out into a vec, then combine the
details of events that took more than one line to
complete.