Remove redundant file collection
Initially the list of files in SyscallData was the only
method to find the files accessed by a PID. When the file_data mod
was added, we now had two ways to finding files, and the file_data
method gives us the access time, which SyscallData cannot.  Removing
the old SyscallData method, as well as removing a list of files opened
from the --stats option. In practice I have not found it useful to have
files listed there since  the addition ofmore accurate data in --pid.