Add check to confirm required strace flags were set
Adds a new check at the start to confirm all needed
data elements are included in the strace output. If
they are not, exit with a message explaining what
needs to be added.

Also updated the file listing to sort within PID by
time accessed, rather than execution length of open.