Support strace's '-ttt' flag
Add support for the `-ttt` flag in strace,
which outputs Unix time with usec accuracy.

Also improved performance by no longer parsing
time stamps for RawData. The timestamps strace
provides can be sorted lexically, and as sorting
is the only reason we read the timestamp, there
is no benefit from fully parsing it.