Time spent in userland added
Capture time spent is userland is now listed. We calculate this by
finding the first and last timestamp for a PID, then subtracting
the time spent performing system tasks. The total time metric has
been updated to include user time as well.

Also, fixed formatting of header on exec list and allow the file
name to be positioned anywhere in args.