Enable testing and builds on unknown-linux-musl
Remove or do not include dependencies that fail to build on Linux
using `musl` as libc.

- `Hashbrown`: Fails to build as one of its dependencies uses a
proc macro. Replaced with `std::collections::HashMap` as the
underlying implementation is now used in the standard library.
- `Jemallocator`: Will not build on the default Rust alpine container
as it does not include `make` or `musl-dev` packages. Using a forked
container I can get `jemalloc` to build, but this even then `rustc`
fails to link to it because it's unable to find the symbol for
`secure_getenv`. Unclear why this occurs as this should be part of
`musl` now and `autotools` found it when building. Will use the
system allocator on `musl`.
- `Criterion`: One of its dependencies uses proc macros. Marked all
dev dependencies as excluding `musl`, which lets tests run. Benches
will not work on `musl`.