Improve thread detection
The existing method of tracking clone calls with `CLONE_THREAD`
captures all new threads spawned during a trace, but will not help
us find any existing threads on a process we start tracing in mid-
stream.

To work around this, the parser now captures the address used for
all `futex` calls made that use a `PRIVATE` flag. Any pid that uses
this address must be in the same thread group / process, allowing us
to deduce which pids are related.