Handle new clone format and improve pid sorting
In older versions of strace, the args to `clone` would be included
in the the start of a split syscall. This was inconsistent with
how other syscalls are handled, but made it easier to parse as the
flags and child pid would be in the same line. More recent versions
have updated `clone` to put its args in the starting message.

To handle this, we now capture the raw events for clone in `PidData`
and iterate through them in pairs to collect the required information
and capture the children / threads of the process.

Also, update sorting by start time to take into account that processes
not forked during the trace could have started at any time, and the time
of first appearance in the trace is essentially random. Now we take all
processes that were not forked and sort by pid. This will not always be
accurate, but it's a decent heuristic. Then the proceses forked during
the trace are sorted by start time and appended.