Don't check futexes on execve pids
Child threads that perform an `execve` will be promoted to a separate
process when this occurs. This change causes any pid that performs an
execve to be removed from the private futex check. This may reduce the
accuracy of that check in some cases, but I don't think it will have a
major impact. The alternative, tracking futexes by time, would be
a lot of work.