Don't check futexes on execve pids
Any thread that invokes `execve` is promoted to its own process.
This change captures this, removing the thread relationship and
adding a parent/child. Currently we assume that any processes or
threads forked by this pid occur after it separates.

Also limit futex checks to processes started prior to the trace.