Capture getpid return value
Use the value returned by `getpid` to improve thread detection.
This will always be the lead pid in the thread group, so we associate
it as a sibling thread. We do not set it as the parent of the calling
pid as the lead thread did not necessarily clone the caller.