Capture values from /proc/<PID>/status
It's common for processes to read from `/proc/<PID>/status` to get
details on themselves or other processes. We can use this to extract
information on pids that were already running prior to the start of
the trace. This will print the rather undescriptive name in `tree`
when no execve was executed, and list it at the top of the `pid` view
as well. The `Tgid` and `Pid` fields are also used to capture thread
relationships.

Because this data may be read for an arbitrary pid, we cannot assume
the proc data pertains to a process we're tracing. We pass the proc
data up to the process hash map once that is generated and then fill
in details on processes captured by the trace.