Use bstr to avoid UTF-8 validation
Reading the entire file into a str at the start causes the full file to
be read and checked as valid UTF-8 on the main thread prior to any other
tasks beginning. This can be a lengthy (hundreds of milliseconds)
process on larger files, which we want to avoid if possible.

To prevent a single-threaded check from blocking, the file is now treated
as a `&[u8]` prior to parsing and selectively validated as UTF-8 during
processing. This allows parsing to begin immediately, spreads validation
across all threads, and skips checking large chunks of the file as valid
UTF-8 altogether. In practice this is a 15-20% performance increase on
my machine.

The downside is that the parsing methods available for `&[u8]` are
significantly less powerful than those on `&str`. We use `bstr` to
regain most of these options, with the major exception of a double-ended
iterator over words.  The `fields_with` method can be used as an equivalent
to `split_ascii_whitespace` other than that it can only iterate forward. We
work around this with `rsplit_str` on ' ', but this may lead to
unexpected failures in parsing if multiple spaces are present.

The next logical step is keep all fields as `&[u8]` until printing, with
the exception of pid and return code, which need to be parsed as
integers.