Convert &str fields to  &[u8]
On large files a significant amount of time is spent validating
that the input it valid UTF-8. In 6f9efc74 we stopped reading the input
file into a str before parsing, which removed a single-threaded
bottleneck at the start of the process.  However, internally we still
validated all parsed fields when reading each line.

This commit converts almost all internal data structures to store parsed
data as `&[u8]`, and skipping most UTF-8 validation as a result. The
exceptions to this, `Execs` and the `BTreeMap` returned by
`directories_opened` perform more complex string parsing and are
operating on very small data sets.

Only when printing do we convert to `&str` and validate, which is a very
small subset of the total data collected. This is a 20-40% performance
increase, varying by size of file. Gains are proportionately greater on
small files, but the absolute time saved on large files is still
significant.