Convert directories to use Path
When inititally testing the new `directories` subcommand, I found that
using `Path` instead of manually splitting the `&str` on `/` and
processing was significantly slower. My assumption at the time was that
this was due to extra UTF-8 validation, but on closer inspection it's
because the split method takes off two layers of path, instead of one as
expected.

This commit converts over to `Path.ancestors()` as this will both
resolve the known bug and be more likely to handle unknown edge cases
with path parsing.

In addition, add tests to confirm results are as expected.