Add additional context to 'directories' output
The initial design of 'directories' included a timestamp, but it was
chosen based on whichever event was processed last, which is
non-deterministic. In addition, the output did not make it clear if the
duration was due to a few very long `open` calls, or many fast ones.

This commit adds three new columns, `first time`, `last time`, and
`count` to the output, and removed `timestamp`. The new time columns
make it easier to see where in the trace relevant events can be found.
The count column shows how often a given directory tree was being accessed,
which may be more useful than duration in many cases.