Cache directory relationships using graph
Fixing the path issues in 7a3ca5f1 revealed that `directory` can be
extremely slow when a pid has many `open`/`openat` calls on long paths.
These are costly because they require an update the the map for each
directory in the path, as well as the cost of parsing the full path
every time.

To mitigate this, we change the update strategy when parsing
`FileEvents` to the following:

1. Create/update a map entry for only the directory being accessed.
2. Cache the path components using a graph. The path is only parsed
   until we find an edge that has already been added to the graph.
   Usually allows us to avoid parsing most of the path.
3. Once all events are parsed, walk the graph and sum all child
   events as we go, creating map entries for any paths that were not
   directly accessed.

The graph is created with its source nodes at the end of the path:

```
/usr/lib/kernel   -> /usr/lib -> /usr -> /
                  /
/usr/lib/python -/
```

This has multiple source nodes, and walking the graph from each will
cause events to be counted multiple times. To get a single source node,
we reverse the graph direction so `/` or the base relative path is the
source and the path ends are the sinks. This allows us to walk the graph
a single time with a post-order depth-first search, which will work its
way down from the leaf of each branch and ensure each edge is walked
only once.