Fix slow futex<->thread mapping
Relating pids using `futex` addresses can be an expensive operation in
traces when many separate addresses are used. Most testing of the parser
has been with Golang programs, which use a relatively limited number of
addresses, so this scenario has not come up. On a ~70 MiB NodeJS trace
parsing was taking several minutes, rather then the expected ~300 ms.
This trace had ~400,000 unique `futex` addresses, making creating and
walking the address graph much more expensive.

In this case none of the addresses were shared between processes, but we
still created a graph with many thousands of edges to and from the same
pid, walking each item and discarding the results, a complete waste of
effort.

To prevent this, we now ensure that an address is referred to by at
least two pids before adding it to the graph.

Closes #5